feat: added type to component and policy - added prop to use component in a more generic way - Created policy to avoid unexpected behaviour when authorizing/denying request

+++ app/Livewire/FriendRequestItem.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Livewire;
 
 use App\Models\FriendRequest;
@@ -10,10 +12,15 @@ use Livewire\Component;
 class FriendRequestItem extends Component
 {
     public FriendRequest $friendRequest;
+
+    /** @var 'sent'|'received' */
+    public string $type;
+
     private string $eventName = 'request-processed';
 
     public function accept(): void
     {
+        $this->authorize('accept', $this->friendRequest);
         $currentUser = Auth::user();
         $userToBefriendId = $currentUser?->id === $this->friendRequest->sender_id ? $this->friendRequest->target_id : $this->friendRequest->sender_id;
         $currentUser?->befriend($userToBefriendId);
@@ -23,6 +30,7 @@ class FriendRequestItem extends Component
 
     public function deny(): void
     {
+        $this->authorize('delete', $this->friendRequest);
         $this->friendRequest->delete();
         $this->dispatch($this->eventName);
     }
